#415: Make Github check name consistent with Sonarqube commercial edition
The check name for decorating Github Pull Requests is currently based on the project name, which prevents setting a global policy for an action having completed on all projects based on a single name. To bring the plugin in-line with the commercial edition, a static name is being used during decoration unless the project is declared as a mono-repo, in which case the project name will be added to the check name.