#1043: Handle Multi Quality Rule metrics during decoration
Sonarqube has the concepts of Multi Quality Rule metrics for quality
gates, which use different metric keys from those used in standard
quality metrics. To ensure we don't cause errors when attempting to
resolve these new rules, if the original search does not find a matching
metric then the mapping between standard and MQR metrics is performed
and, if a match is found, the mapped key used to retrieve a metric. If
no mapping is found then the current quality gate condition is skipped
during the decoration.