Update README to reflect installation of webapp components
Adds webapp unpacking instructions to the manual install, and an
extra volume, initContainer, and volumeMount to download and mount the
webapp contents in place of the bundled webapp in Helm. Includes the
removal of references to workarounds for old versions of the
Sonarqube Helm chart since this is no longer supported.