#1075: Switch classloader injection to reference a stable class
The class reference class loader strategy currently loads a class from
Sonarqube core that will be removed in future versions of Sonarqube
since the Profile import and export features this class is part of have
been deprecated. The reference is being replaced a non-deprecated class
from the same API component.