#1049: Rebuild the web interface to include branch features
Sonarqube has had most of the branch features removed from the community edition's Web UI, resulting in the branch selection menu being disabled even where branches or pull requests are available. To overcome this, the sonarqube-web project is being built as part of the plugin build, with the missing branch components re-instated and the configuration set to enable branch functionality. The build and release process have been updated to publish the web ZIP alongside the plugin JAR, and the Readme updated with guidance for how to launch a local and Helm instance with the updated web interface.