#920: Add guidance on sonar.scm.revision in Github CI PR builds
As come CI build, such as Gitlab CI, perform a merge commit when checking out a Pull Request for building, the commit in the local workspace may not match the commit hash in the Pull Request, so Sonarqube doesn't find the commit to attach decoration comments to. To overcome this, guidance is being added on using the `sonar.scm.revision` property alongside the relevant Gitlab CI properties to pickup the PR's source commit hash.