#1005: Use the correct target branch ID when scanning a new project
During the first analysis of a non-main branch on a new project, the plugin currently set the branch's reference branch as the project ID. However, this results in the analysis failing since the target branch can't be found, so the Sonarqube UI reports the branch hasn't been analysed. To overcome this, the ID resolution is being altered to lookup the ID of the default 'main' branch on the project and use that as the reference ID for the current branch, thereby ensuring that the analysis can execute.