#1078: Ensure we're building the customised webapp
Older versions of sonarqube-webapp have the additional components in an
`addons` rather than `sq-server-addons` directory, so the build
customisation isn't creating a modified front-end. The project setup has
therefore been updated to use the path the webapp expects, and the
branch components altered to use the right imports for the target
version of Sonarqube. To simplify the plugin versioning, the plugin
version has been altered to match the major and minor version of the
Sonarqube distribution it's aiming to support.