#897: Truncate report keys when sending to Bitbucket server
When using a monorepo with a long project key, Bitbucket server reports
an error for the report key being too long during decoration. As the
report key is based on the project key which is typically derived from a
project's group ID and artefact ID, this may end up being a long value
with limited control over setting its length. To overcome this, when
sending a report to Bitbucket server, the project key will be truncated
where it's more than 50 characters, and prepended with `t-` to make it
clear the value was truncated and try and prevent conflicts with any
other projects.