#1133: Add support for disabling inline comments in Azure decoration
Sonarqube provides a toggle in the front-end to disable inline comments
on Azure DevOps repository decoration. This parameter was being ignored
in the endpoint handling, and the decorator did not check for the
value in the project settings, so the configured value was ignored and
inline comments were treated as always being enabled. The Azure bindings
endpoints has been altered to parse the `inlineAnnotationsEnabled`
parameter and save it in the project settings, and the Azure decorator
modified to check this parameter before performing any operations on
inline comments in a Pull request.