#249: Add Atlassian XSRF protection to Bitbucket Server request
Bitbucket server performs an XSRF check that requires the content-type
is either complex (e.g. application/json) or that a special request
header is added that indicates the request has been sent from a trusted
source. As no content-type is set on GET or DELETE requests, the
Atlassian XSRF header is being set on the request to indicate the
request has come from a trusted source rather than being a cross-site
request from an authenticated front-end.