#1115: Differentiate Gitlab pipelines on monorepos
The Gitlab decorator used a constant name for the pipeline which meant
the last executed pipeline in a monorepo may allow a Merge Request to be
marked as successful even if other projects in the repository had failed
their pipelines. To overcome this the decorator has been altered to
include the project key in the pipeline name if the project is a
monorepo to ensure that the pipeline is unique across all projects in
the monorepo.